+++ mathics/builtin/lists.py
@@ -5,7 +5,7 @@ List Functions - Miscellaneous
 Functions here will eventually get moved to more suitable subsections.
 """
 
-from mathics.algorithm.parts import python_levelspec, walk_levels
+from mathics.algorithm.parts import python_levelspec
 from mathics.builtin.base import Builtin, Predefined, Test
 from mathics.builtin.box.layout import RowBox
 from mathics.core.atoms import Integer, Integer1, Integer2, String
@@ -362,154 +362,6 @@ class IntersectingQ(Builtin):
 
     rules = {"IntersectingQ[a_List, b_List]": "Length[Intersect[a, b]] > 0"}
     summary_text = "test whether two lists have common elements"
-
-
-class LeafCount(Builtin):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/LeafCount.html</url>
-
-    <dl>
-      <dt>'LeafCount[$expr$]'
-      <dd>returns the total number of indivisible subexpressions in $expr$.
-    </dl>
-
-    >> LeafCount[1 + x + y^a]
-     = 6
-
-    >> LeafCount[f[x, y]]
-     = 3
-
-    >> LeafCount[{1 / 3, 1 + I}]
-     = 7
-
-    >> LeafCount[Sqrt[2]]
-     = 5
-
-    >> LeafCount[100!]
-     = 1
-
-    #> LeafCount[f[a, b][x, y]]
-     = 5
-
-    #> NestList[# /. s[x_][y_][z_] -> x[z][y[z]] &, s[s][s][s[s]][s][s], 4];
-    #> LeafCount /@ %
-     = {7, 8, 8, 11, 11}
-
-    #> LeafCount[1 / 3, 1 + I]
-     : LeafCount called with 2 arguments; 1 argument is expected.
-     = LeafCount[1 / 3, 1 + I]
-    """
-
-    messages = {
-        "argx": "LeafCount called with `1` arguments; 1 argument is expected.",
-    }
-    summary_text = "the total number of atomic subexpressions"
-
-    def eval(self, expr, evaluation):
-        "LeafCount[expr___]"
-
-        from mathics.core.atoms import Complex, Rational
-
-        elements = []
-
-        def callback(level):
-            if isinstance(level, Rational):
-                elements.extend(
-                    [level.get_head(), level.numerator(), level.denominator()]
-                )
-            elif isinstance(level, Complex):
-                elements.extend([level.get_head(), level.real, level.imag])
-            else:
-                elements.append(level)
-            return level
-
-        expr = expr.get_sequence()
-        if len(expr) != 1:
-            return evaluation.message("LeafCount", "argx", Integer(len(expr)))
-
-        walk_levels(expr[0], start=-1, stop=-1, heads=True, callback=callback)
-        return Integer(len(elements))
-
-
-class Level(Builtin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/Level.html</url>
-
-    <dl>
-      <dt>'Level[$expr$, $levelspec$]'
-      <dd>gives a list of all subexpressions of $expr$ at the
-        level(s) specified by $levelspec$.
-    </dl>
-
-    Level uses standard level specifications:
-
-    <dl>
-      <dt>$n$
-      <dd>levels 1 through $n$
-      <dt>'Infinity'
-      <dd>all levels from level 1
-      <dt>'{$n$}'
-      <dd>level $n$ only
-      <dt>'{$m$, $n$}'
-      <dd>levels $m$ through $n$
-    </dl>
-
-    Level 0 corresponds to the whole expression.
-
-    A negative level '-$n$' consists of parts with depth $n$.
-
-    Level -1 is the set of atoms in an expression:
-    >> Level[a + b ^ 3 * f[2 x ^ 2], {-1}]
-     = {a, b, 3, 2, x, 2}
-
-    >> Level[{{{{a}}}}, 3]
-     = {{a}, {{a}}, {{{a}}}}
-    >> Level[{{{{a}}}}, -4]
-     = {{{{a}}}}
-    >> Level[{{{{a}}}}, -5]
-     = {}
-
-    >> Level[h0[h1[h2[h3[a]]]], {0, -1}]
-     = {a, h3[a], h2[h3[a]], h1[h2[h3[a]]], h0[h1[h2[h3[a]]]]}
-
-    Use the option 'Heads -> True' to include heads:
-    >> Level[{{{{a}}}}, 3, Heads -> True]
-     = {List, List, List, {a}, {{a}}, {{{a}}}}
-    >> Level[x^2 + y^3, 3, Heads -> True]
-     = {Plus, Power, x, 2, x ^ 2, Power, y, 3, y ^ 3}
-
-    >> Level[a ^ 2 + 2 * b, {-1}, Heads -> True]
-     = {Plus, Power, a, 2, Times, 2, b}
-    >> Level[f[g[h]][x], {-1}, Heads -> True]
-     = {f, g, h, x}
-    >> Level[f[g[h]][x], {-2, -1}, Heads -> True]
-     = {f, g, h, g[h], x, f[g[h]][x]}
-    """
-
-    options = {
-        "Heads": "False",
-    }
-    summary_text = "parts specified by a given number of indices"
-
-    def eval(self, expr, ls, evaluation, options={}):
-        "Level[expr_, ls_, OptionsPattern[Level]]"
-
-        try:
-            start, stop = python_levelspec(ls)
-        except InvalidLevelspecError:
-            evaluation.message("Level", "level", ls)
-            return
-        result = []
-
-        def callback(level):
-            result.append(level)
-            return level
-
-        heads = self.get_option(options, "Heads", evaluation) is SymbolTrue
-        walk_levels(expr, start, stop, heads=heads, callback=callback)
-        return ListExpression(*result)
 
 
 class LevelQ(Test):
